Jump to beta! statement parser works, reasonably debugged and tested. This version seems functional in respect to what was planned. Next milestone: check & fix remaining memory leaks, add documentation.

+++ trunk/ninjac/builtins.h
@@ -19,7 +19,7 @@ public:
 
 class BuiltInRand : public Expression {
 public:
-    virtual double evaluate () const { return (double)rand() / (double)RAND_MAX; }
+    virtual double evaluate () const { return (double)rand() / RAND_MAX; }
 };
 
 class BuiltInNot : public Expression {
